test(hnsw): move coverage tests to unit test suite
